some follow ups for getallmodels adoption
