+++ packages/http-client-csharp/emitter/src/lib/converter.ts
@@ -24,20 +24,18 @@ import {
 import {
   DateTimeKnownEncoding,
   DurationKnownEncoding,
+  EncodeData,
   IntrinsicType,
   Model,
   Program,
   Type,
   getFormat,
 } from "@typespec/compiler";
-import { logger } from "../lib/logger.js";
-import { getCSharpInputTypeKindByPrimitiveModelName } from "../lib/model.js";
-import { getFullNamespaceString } from "../lib/utils.js";
-import { InputEnumTypeValue } from "./input-enum-type-value.js";
-import { InputIntrinsicTypeKind } from "./input-intrinsic-type-kind.js";
-import { InputModelProperty } from "./input-model-property.js";
-import { InputPrimitiveTypeKind } from "./input-primitive-type-kind.js";
-import { InputTypeKind } from "./input-type-kind.js";
+import { InputEnumTypeValue } from "../type/input-enum-type-value.js";
+import { InputIntrinsicTypeKind } from "../type/input-intrinsic-type-kind.js";
+import { InputModelProperty } from "../type/input-model-property.js";
+import { InputPrimitiveTypeKind } from "../type/input-primitive-type-kind.js";
+import { InputTypeKind } from "../type/input-type-kind.js";
 import {
   InputDictionaryType,
   InputEnumType,
@@ -48,11 +46,13 @@ import {
   InputPrimitiveType,
   InputType,
   InputUnionType,
-} from "./input-type.js";
-import { LiteralTypeContext } from "./literal-type-context.js";
-import { Usage } from "./usage.js";
+} from "../type/input-type.js";
+import { LiteralTypeContext } from "../type/literal-type-context.js";
+import { Usage } from "../type/usage.js";
+import { logger } from "./logger.js";
+import { getFullNamespaceString } from "./utils.js";
 
-function fromSdkType(
+export function fromSdkType(
   sdkType: SdkType,
   context: SdkContext,
   models: Map<string, InputModelType>,
@@ -79,6 +79,22 @@ function fromSdkType(
   if (sdkType.__raw?.kind === "Intrinsic") return fromIntrinsicType(sdkType);
   if (isSdkBuiltInKind(sdkType.kind)) return fromSdkBuiltInType(sdkType as SdkBuiltInType);
   return {} as InputType;
+}
+
+// TODO -- this is workaround because TCGC ignore format, we need to remove this after a discussion on format.
+// this function is only for the case when we get a type from a parameter, because in typespec, the parameters share the same type as the properties
+export function fromSdkModelPropertyType(
+  propertyType: SdkModelPropertyType,
+  context: SdkContext,
+  models: Map<string, InputModelType>,
+  enums: Map<string, InputEnumType>,
+  literalTypeContext?: LiteralTypeContext
+): InputType {
+  // when the type is string, we need to add the format
+  if (propertyType.type.kind === "string") {
+    return fromStringType(context.program, propertyType.type, propertyType.__raw);
+  }
+  return fromSdkType(propertyType.type, context, models, enums, literalTypeContext);
 }
 
 export function fromSdkModelType(
@@ -129,7 +145,7 @@ export function fromSdkModelType(
       .filter((p) => !(p as SdkBodyModelPropertyType).discriminator || !baseModelHasDiscriminator)
       .filter((p) => p.kind !== "header" && p.kind !== "query" && p.kind !== "path")
       .map((p) =>
-        fromSdkModelPropertyType(p, context, models, enums, {
+        fromSdkModelProperty(p, {
           ModelName: inputModelType?.Name,
           Namespace: inputModelType?.Namespace,
         } as LiteralTypeContext)
@@ -137,6 +153,33 @@ export function fromSdkModelType(
   }
 
   return inputModelType;
+
+  function fromSdkModelProperty(
+    propertyType: SdkModelPropertyType,
+    literalTypeContext: LiteralTypeContext
+  ): InputModelProperty {
+    const serializedName =
+      propertyType.kind === "property"
+        ? (propertyType as SdkBodyModelPropertyType).serializedName
+        : "";
+    literalTypeContext.PropertyName = serializedName;
+
+    const isRequired =
+      propertyType.kind === "path" || propertyType.kind === "body" ? true : !propertyType.optional; // TO-DO: SdkBodyParameter lacks of optional
+    const isDiscriminator =
+      propertyType.kind === "property" && propertyType.discriminator ? true : false;
+    const modelProperty: InputModelProperty = {
+      Name: propertyType.name,
+      SerializedName: serializedName,
+      Description: propertyType.description ?? (isDiscriminator ? "Discriminator" : ""),
+      Type: fromSdkType(propertyType.type, context, models, enums, literalTypeContext),
+      IsRequired: isRequired,
+      IsReadOnly: propertyType.kind === "property" && isReadOnly(propertyType),
+      IsDiscriminator: isDiscriminator === true ? true : undefined, // TODO: keep backward compatible to ease comparison. remove this after TCGC is merged
+    };
+
+    return modelProperty;
+  }
 }
 
 function getDiscriminatorPropertyNameFromCurrentModel(model?: SdkModelType): string | undefined {
@@ -164,7 +207,7 @@ export function fromSdkEnumType(
   enums: Map<string, InputEnumType>,
   addToCollection: boolean = true
 ): InputEnumType {
-  const enumName = enumType.name;
+  let enumName = enumType.name;
   let inputEnumType = enums.get(enumName);
   if (inputEnumType === undefined) {
     const newInputEnumType: InputEnumType = {
@@ -221,6 +264,9 @@ function fromSdkDurationType(durationType: SdkDurationType): InputPrimitiveType 
         if (wireType.kind === "float" || wireType.kind === "float32") {
           return InputPrimitiveTypeKind.DurationSecondsFloat;
         }
+        if (wireType.kind === "float64") {
+          return InputPrimitiveTypeKind.DurationSecondsDouble;
+        }
         return InputPrimitiveTypeKind.DurationSeconds;
       default:
         logger.warn(`invalid encode '${encode}' and wireType '${wireType.kind}' for duration.`);
@@ -265,7 +311,14 @@ function fromBytesType(bytesType: SdkBuiltInType): InputPrimitiveType {
   };
 }
 
-function fromStringType(program: Program, stringType: SdkType): InputPrimitiveType {
+function fromStringType(
+  program: Program,
+  stringType: SdkType,
+  // we need the extra raw here because the format decorator is added to the property/parameter, but the raw in stringType is the type itself, and it does not have the format decorator we want.
+  // only when we get the type from a parameter, we need to pass the the parameter as raw here to get the format
+  // TODO -- we should remove this entirely later because TCGC ignores format in these cases, we add it now because we have old test projects, and we did not discuss the impact yet
+  raw?: Type
+): InputPrimitiveType {
   function fromStringFormat(rawStringType?: Type): InputPrimitiveTypeKind {
     if (!rawStringType) return InputPrimitiveTypeKind.String;
 
@@ -287,9 +340,10 @@ function fromStringType(program: Program, stringType: SdkType): InputPrimitiveTy
     }
   }
 
+  raw = raw ?? stringType.__raw;
   return {
     Kind: InputTypeKind.Primitive,
-    Name: fromStringFormat(stringType.__raw),
+    Name: fromStringFormat(raw),
     IsNullable: stringType.nullable,
   };
 }
@@ -386,9 +440,116 @@ function fromScalarType(scalarType: SdkType): InputPrimitiveType {
     ),
     IsNullable: scalarType.nullable,
   };
+
+  function getCSharpInputTypeKindByPrimitiveModelName(
+    name: string,
+    format?: string,
+    encode?: EncodeData
+  ): InputPrimitiveTypeKind {
+    switch (name) {
+      case "bytes":
+        switch (encode?.encoding) {
+          case undefined:
+          case "base64":
+            return InputPrimitiveTypeKind.Bytes;
+          case "base64url":
+            return InputPrimitiveTypeKind.BytesBase64Url;
+          default:
+            logger.warn(`invalid encode ${encode?.encoding} for bytes.`);
+            return InputPrimitiveTypeKind.Bytes;
+        }
+      case "int8":
+        return InputPrimitiveTypeKind.SByte;
+      case "uint8":
+        return InputPrimitiveTypeKind.Byte;
+      case "int32":
+        return InputPrimitiveTypeKind.Int32;
+      case "int64":
+        return InputPrimitiveTypeKind.Int64;
+      case "integer":
+        return InputPrimitiveTypeKind.Int64;
+      case "safeint":
+        return InputPrimitiveTypeKind.SafeInt;
+      case "float32":
+        return InputPrimitiveTypeKind.Float32;
+      case "float64":
+        return InputPrimitiveTypeKind.Float64;
+      case "decimal":
+        return InputPrimitiveTypeKind.Decimal;
+      case "decimal128":
+        return InputPrimitiveTypeKind.Decimal128;
+      case "uri":
+      case "url":
+        return InputPrimitiveTypeKind.Uri;
+      case "uuid":
+        return InputPrimitiveTypeKind.Guid;
+      case "eTag":
+        return InputPrimitiveTypeKind.String;
+      case "string":
+        switch (format?.toLowerCase()) {
+          case "date":
+            return InputPrimitiveTypeKind.DateTime;
+          case "uri":
+          case "url":
+            return InputPrimitiveTypeKind.Uri;
+          case "uuid":
+            return InputPrimitiveTypeKind.Guid;
+          default:
+            if (format) {
+              logger.warn(`invalid format ${format}`);
+            }
+            return InputPrimitiveTypeKind.String;
+        }
+      case "boolean":
+        return InputPrimitiveTypeKind.Boolean;
+      case "date":
+        return InputPrimitiveTypeKind.Date;
+      case "plainDate":
+        return InputPrimitiveTypeKind.Date;
+      case "plainTime":
+        return InputPrimitiveTypeKind.Time;
+      case "datetime":
+      case "utcDateTime":
+        switch (encode?.encoding) {
+          case undefined:
+            return InputPrimitiveTypeKind.DateTime;
+          case "rfc3339":
+            return InputPrimitiveTypeKind.DateTimeRFC3339;
+          case "rfc7231":
+            return InputPrimitiveTypeKind.DateTimeRFC7231;
+          case "unixTimestamp":
+            return InputPrimitiveTypeKind.DateTimeUnix;
+          default:
+            logger.warn(`invalid encode ${encode?.encoding} for date time.`);
+            return InputPrimitiveTypeKind.DateTime;
+        }
+      case "time":
+        return InputPrimitiveTypeKind.Time;
+      case "duration":
+        switch (encode?.encoding) {
+          case undefined:
+          case "ISO8601":
+            return InputPrimitiveTypeKind.DurationISO8601;
+          case "seconds":
+            if (encode.type?.name === "float" || encode.type?.name === "float32") {
+              return InputPrimitiveTypeKind.DurationSecondsFloat;
+            } else {
+              return InputPrimitiveTypeKind.DurationSeconds;
+            }
+          default:
+            logger.warn(`invalid encode ${encode?.encoding} for duration.`);
+            return InputPrimitiveTypeKind.DurationISO8601;
+        }
+      case "azureLocation":
+        return InputPrimitiveTypeKind.AzureLocation;
+      default:
+        return InputPrimitiveTypeKind.Object;
+    }
+  }
 }
 
 function fromIntrinsicType(scalarType: SdkType): InputIntrinsicType {
+  const name = (scalarType.__raw! as IntrinsicType).name;
   return {
     Kind: InputTypeKind.Intrinsic,
     Name: getCSharpInputTypeKindByIntrinsic(scalarType.__raw! as IntrinsicType),
@@ -402,7 +563,7 @@ function fromUnionType(
   models: Map<string, InputModelType>,
   enums: Map<string, InputEnumType>
 ): InputUnionType | InputType {
-  const itemTypes: InputType[] = [];
+  let itemTypes: InputType[] = [];
   for (const value of union.values) {
     const inputType = fromSdkType(value, context, models, enums);
     itemTypes.push(inputType);
@@ -534,36 +695,6 @@ function fromUsageFlags(usage: UsageFlags): Usage {
   else if (usage === UsageFlags.Output) return Usage.Output;
   else if (usage === (UsageFlags.Input | UsageFlags.Output)) return Usage.RoundTrip;
   else return Usage.None;
-}
-
-function fromSdkModelPropertyType(
-  propertyType: SdkModelPropertyType,
-  context: SdkContext,
-  models: Map<string, InputModelType>,
-  enums: Map<string, InputEnumType>,
-  literalTypeContext: LiteralTypeContext
-): InputModelProperty {
-  const serializedName =
-    propertyType.kind === "property"
-      ? (propertyType as SdkBodyModelPropertyType).serializedName
-      : "";
-  literalTypeContext.PropertyName = serializedName;
-
-  const isRequired =
-    propertyType.kind === "path" || propertyType.kind === "body" ? true : !propertyType.optional; // TO-DO: SdkBodyParameter lacks of optional
-  const isDiscriminator =
-    propertyType.kind === "property" && propertyType.discriminator ? true : false;
-  const modelProperty: InputModelProperty = {
-    Name: propertyType.nameInClient,
-    SerializedName: serializedName,
-    Description: propertyType.description ?? (isDiscriminator ? "Discriminator" : ""),
-    Type: fromSdkType(propertyType.type, context, models, enums, literalTypeContext),
-    IsRequired: isRequired,
-    IsReadOnly: propertyType.kind === "property" && isReadOnly(propertyType),
-    IsDiscriminator: isDiscriminator === true ? true : undefined, // TODO: keep backward compatible to ease comparison. remove this after TCGC is merged
-  };
-
-  return modelProperty;
 }
 
 function getCSharpInputTypeKindByIntrinsic(type: IntrinsicType): InputIntrinsicTypeKind {
