+++ packages/http-client-csharp/emitter/src/lib/model.ts
@@ -1,33 +1,21 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { getLroMetadata, isFixed } from "@azure-tools/typespec-azure-core";
+import { getLroMetadata } from "@azure-tools/typespec-azure-core";
 import {
   SdkContext,
-  SdkEnumType,
-  SdkModelType,
   getAllModels,
   getClientType,
+  getSdkModelPropertyType,
 } from "@azure-tools/typespec-client-generator-core";
 import {
-  EncodeData,
-  Enum,
-  IntrinsicType,
   Model,
   ModelProperty,
   Operation,
-  Program,
-  ProjectedProgram,
-  Scalar,
   Type,
-  Union,
   UsageFlags,
-  getDeprecated,
-  getDoc,
   getEffectiveModelType,
-  getEncode,
-  getFormat,
-  getKnownValues,
+  ignoreDiagnostics,
   isArrayModelType,
   isRecordModelType,
   resolveUsages,
@@ -39,168 +27,24 @@ import {
   getQueryParamName,
   isStatusCode,
 } from "@typespec/http";
+import { getResourceOperation } from "@typespec/rest";
 import { NetEmitterOptions } from "../options.js";
-import { fromSdkEnumType, fromSdkModelType } from "../type/converter.js";
-import { FormattedType } from "../type/formatted-type.js";
-import { InputEnumTypeValue } from "../type/input-enum-type-value.js";
-import { InputIntrinsicTypeKind } from "../type/input-intrinsic-type-kind.js";
-import { InputPrimitiveTypeKind } from "../type/input-primitive-type-kind.js";
-import { InputTypeKind } from "../type/input-type-kind.js";
 import {
-  InputDictionaryType,
   InputEnumType,
-  InputIntrinsicType,
-  InputListType,
-  InputLiteralType,
   InputModelType,
-  InputPrimitiveType,
   InputType,
-  InputUnionType,
   isInputEnumType,
-  isInputIntrinsicType,
   isInputLiteralType,
 } from "../type/input-type.js";
 import { LiteralTypeContext } from "../type/literal-type-context.js";
+import {
+  fromSdkEnumType,
+  fromSdkModelPropertyType,
+  fromSdkModelType,
+  fromSdkType,
+} from "./converter.js";
 import { logger } from "./logger.js";
-import { capitalize, getFullNamespaceString, getTypeName } from "./utils.js";
-/**
- * Map calType to csharp InputTypeKind
- */
-export function mapTypeSpecTypeToCSharpInputTypeKind(
-  typespecType: Type,
-  format?: string,
-  encode?: EncodeData
-): InputPrimitiveTypeKind {
-  const kind = typespecType.kind;
-  switch (kind) {
-    case "Model":
-      return getCSharpInputTypeKindByPrimitiveModelName(typespecType.name, format, encode);
-    case "ModelProperty":
-      return InputPrimitiveTypeKind.Object;
-    case "Enum":
-      return InputPrimitiveTypeKind.Enum;
-    case "Number":
-      const numberValue = typespecType.value;
-      if (numberValue % 1 === 0) {
-        return InputPrimitiveTypeKind.Int32;
-      }
-      return InputPrimitiveTypeKind.Float64;
-    case "Boolean":
-      return InputPrimitiveTypeKind.Boolean;
-    case "String":
-      if (format === "date") return InputPrimitiveTypeKind.DateTime;
-      if (format === "uri") return InputPrimitiveTypeKind.Uri;
-      return InputPrimitiveTypeKind.String;
-    default:
-      throw new Error(`Unsupported primitive kind ${kind}`);
-  }
-}
-
-export function getCSharpInputTypeKindByPrimitiveModelName(
-  name: string,
-  format?: string,
-  encode?: EncodeData
-): InputPrimitiveTypeKind {
-  switch (name) {
-    case "bytes":
-      switch (encode?.encoding) {
-        case undefined:
-        case "base64":
-          return InputPrimitiveTypeKind.Bytes;
-        case "base64url":
-          return InputPrimitiveTypeKind.BytesBase64Url;
-        default:
-          logger.warn(`invalid encode ${encode?.encoding} for bytes.`);
-          return InputPrimitiveTypeKind.Bytes;
-      }
-    case "int8":
-      return InputPrimitiveTypeKind.SByte;
-    case "uint8":
-      return InputPrimitiveTypeKind.Byte;
-    case "int32":
-      return InputPrimitiveTypeKind.Int32;
-    case "int64":
-      return InputPrimitiveTypeKind.Int64;
-    case "integer":
-      return InputPrimitiveTypeKind.Int64;
-    case "safeint":
-      return InputPrimitiveTypeKind.SafeInt;
-    case "float32":
-      return InputPrimitiveTypeKind.Float32;
-    case "float64":
-      return InputPrimitiveTypeKind.Float64;
-    case "decimal":
-      return InputPrimitiveTypeKind.Decimal;
-    case "decimal128":
-      return InputPrimitiveTypeKind.Decimal128;
-    case "uri":
-    case "url":
-      return InputPrimitiveTypeKind.Uri;
-    case "uuid":
-      return InputPrimitiveTypeKind.Guid;
-    case "eTag":
-      return InputPrimitiveTypeKind.String;
-    case "string":
-      switch (format?.toLowerCase()) {
-        case "date":
-          return InputPrimitiveTypeKind.DateTime;
-        case "uri":
-        case "url":
-          return InputPrimitiveTypeKind.Uri;
-        case "uuid":
-          return InputPrimitiveTypeKind.Guid;
-        default:
-          if (format) {
-            logger.warn(`invalid format ${format}`);
-          }
-          return InputPrimitiveTypeKind.String;
-      }
-    case "boolean":
-      return InputPrimitiveTypeKind.Boolean;
-    case "date":
-      return InputPrimitiveTypeKind.Date;
-    case "plainDate":
-      return InputPrimitiveTypeKind.Date;
-    case "plainTime":
-      return InputPrimitiveTypeKind.Time;
-    case "datetime":
-    case "utcDateTime":
-      switch (encode?.encoding) {
-        case undefined:
-          return InputPrimitiveTypeKind.DateTime;
-        case "rfc3339":
-          return InputPrimitiveTypeKind.DateTimeRFC3339;
-        case "rfc7231":
-          return InputPrimitiveTypeKind.DateTimeRFC7231;
-        case "unixTimestamp":
-          return InputPrimitiveTypeKind.DateTimeUnix;
-        default:
-          logger.warn(`invalid encode ${encode?.encoding} for date time.`);
-          return InputPrimitiveTypeKind.DateTime;
-      }
-    case "time":
-      return InputPrimitiveTypeKind.Time;
-    case "duration":
-      switch (encode?.encoding) {
-        case undefined:
-        case "ISO8601":
-          return InputPrimitiveTypeKind.DurationISO8601;
-        case "seconds":
-          if (encode.type?.name === "float" || encode.type?.name === "float32") {
-            return InputPrimitiveTypeKind.DurationSecondsFloat;
-          } else {
-            return InputPrimitiveTypeKind.DurationSeconds;
-          }
-        default:
-          logger.warn(`invalid encode ${encode?.encoding} for duration.`);
-          return InputPrimitiveTypeKind.DurationISO8601;
-      }
-    case "azureLocation":
-      return InputPrimitiveTypeKind.AzureLocation;
-    default:
-      return InputPrimitiveTypeKind.Object;
-  }
-}
+import { capitalize, getTypeName } from "./utils.js";
 
 /**
  * If type is an anonymous model, tries to find a named model that has the same
@@ -254,296 +98,24 @@ export function getDefaultValue(type: Type): any {
 
 export function getInputType(
   context: SdkContext<NetEmitterOptions>,
-  formattedType: FormattedType,
+  type: Type,
   models: Map<string, InputModelType>,
   enums: Map<string, InputEnumType>,
+  operation?: Operation,
   literalTypeContext?: LiteralTypeContext
 ): InputType {
-  const type = getRealType(
-    formattedType.type.kind === "ModelProperty" ? formattedType.type.type : formattedType.type,
-    context
-  );
   logger.debug(`getInputType for kind: ${type.kind}`);
-  const program = context.program;
 
-  if (type.kind === "Model") {
-    return getInputModelType(type);
-  } else if (type.kind === "String" || type.kind === "Number" || type.kind === "Boolean") {
-    return getInputLiteralType(formattedType, literalTypeContext);
-  } else if (type.kind === "Enum") {
-    return getInputTypeForEnum(type);
-  } else if (type.kind === "EnumMember") {
-    return getInputTypeForEnum(type.enum);
-  } else if (type.kind === "Intrinsic") {
-    return getInputModelForIntrinsicType(type);
-  } else if (type.kind === "Scalar") {
-    let effectiveType = type;
-    while (!program.checker.isStdType(effectiveType)) {
-      if (type.baseScalar) {
-        effectiveType = type.baseScalar;
-      } else {
-        break;
-      }
-    }
-    const intrinsicName = effectiveType.name;
-    switch (intrinsicName) {
-      case "string":
-        const values = getKnownValues(program, type);
-        if (values) {
-          return getInputModelForEnumByKnowValues(type, values);
-        }
-      // if the model is one of the typespec Intrinsic type.
-      // it's a base typespec "primitive" that corresponds directly to an c# data type.
-      // In such cases, we don't want to emit a ref and instead just
-      // emit the base type directly.
-      // TODO: verify this is good might be a bug
-      // eslint-disable-next-line no-fallthrough
-      default:
-        const sdkType = getClientType(context, type);
-        return {
-          Kind: InputTypeKind.Primitive,
-          Name: getCSharpInputTypeKindByPrimitiveModelName(
-            sdkType.kind,
-            formattedType.format,
-            formattedType.encode
-          ),
-          IsNullable: false,
-        } as InputPrimitiveType;
-    }
-  } else if (type.kind === "Union") {
-    return getInputTypeForUnion(type);
-  } else if (type.kind === "UnionVariant") {
-    return getInputType(
-      context,
-      getFormattedType(program, type.type),
-      models,
-      enums,
-      literalTypeContext
-    );
-  } else if (type.kind === "Tuple") {
-    return {
-      Kind: InputTypeKind.Intrinsic,
-      Name: InputIntrinsicTypeKind.Unknown,
-      IsNullable: false,
-    } as InputIntrinsicType;
-  } else {
-    throw new Error(`Unsupported type ${type.kind}`);
+  // TODO -- we might could remove this workaround when we adopt getAllOperations
+  //         or when we decide not to honor the `@format` decorators on parameters
+  // this is specifically dealing with the case of an operation parameter
+  if (type.kind === "ModelProperty") {
+    const propertyType = ignoreDiagnostics(getSdkModelPropertyType(context, type, operation));
+    return fromSdkModelPropertyType(propertyType, context, models, enums, literalTypeContext);
   }
 
-  function getInputModelType(m: Model): InputListType | InputDictionaryType | InputModelType {
-    /* Array and Map Type. */
-    if (isArrayModelType(program, m)) {
-      return getInputTypeForArray(m.indexer.value);
-    } else if (isRecordModelType(program, m) && m.sourceModel === undefined) {
-      // only when the model does not have a source model, it is really a record type
-      // when we have `model Foo is Record<string>` this should be a model with additional properties therefore it should not be parsed into a dictionary type
-      return getInputTypeForMap(m.indexer.key, m.indexer.value);
-    }
-    return getInputModelForModel(m);
-  }
-
-  function getInputModelForEnumByKnowValues(m: Model | Scalar, e: Enum): InputEnumType {
-    const name = getTypeName(context, m);
-    let extensibleEnum = enums.get(name);
-    if (!extensibleEnum) {
-      const innerEnum: InputEnumType = getInputTypeForEnum(e, false);
-      if (!innerEnum) {
-        throw new Error(`Extensible enum type '${e.name}' has no values defined.`);
-      }
-      extensibleEnum = {
-        Name: name,
-        EnumValueType: innerEnum.EnumValueType, //EnumValueType and  AllowedValues should be the first field after id and name, so that it can be corrected serialized.
-        AllowedValues: innerEnum.AllowedValues,
-        Namespace: getFullNamespaceString(e.namespace),
-        Accessibility: undefined, //TODO: need to add accessibility
-        Deprecated: getDeprecated(program, m),
-        Description: getDoc(program, m),
-        IsExtensible: !isFixed(program, e),
-        IsNullable: false,
-      } as InputEnumType;
-      enums.set(name, extensibleEnum);
-    }
-    return extensibleEnum;
-  }
-
-  function getInputLiteralType(
-    formattedType: FormattedType,
-    literalContext?: LiteralTypeContext
-  ): InputLiteralType {
-    // For literal types, we just want to emit them directly as well.
-    const type = formattedType.type;
-    const builtInKind: InputPrimitiveTypeKind = mapTypeSpecTypeToCSharpInputTypeKind(
-      type,
-      formattedType.format,
-      formattedType.encode
-    );
-    const rawValueType: InputPrimitiveType = {
-      Kind: InputTypeKind.Primitive,
-      Name: builtInKind,
-      IsNullable: false,
-    };
-    const literalValue = getDefaultValue(type);
-    const newValueType = getLiteralValueType();
-
-    if (isInputEnumType(newValueType)) {
-      enums.set(newValueType.Name, newValueType);
-    }
-
-    return {
-      Kind: InputTypeKind.Literal,
-      Name: InputTypeKind.Literal,
-      LiteralValueType: newValueType,
-      Value: literalValue,
-      IsNullable: false,
-    };
-
-    function getLiteralValueType(): InputPrimitiveType | InputEnumType {
-      // we will not wrap it if it comes from outside a model or it is a boolean
-      if (literalContext === undefined || rawValueType.Name === InputPrimitiveTypeKind.Boolean)
-        return rawValueType;
-
-      // otherwise we need to wrap this into an extensible enum
-      // we use the model name followed by the property name as the enum name to ensure it is unique
-      const enumName = `${literalContext.ModelName}_${literalContext.PropertyName}`;
-      const enumValueType =
-        rawValueType.Name === InputPrimitiveTypeKind.String
-          ? InputPrimitiveTypeKind.String
-          : InputPrimitiveTypeKind.Float32;
-      const allowValues: InputEnumTypeValue[] = [
-        {
-          Name: literalValue.toString(),
-          Value: literalValue,
-          Description: literalValue.toString(),
-        },
-      ];
-      const enumType: InputEnumType = {
-        Kind: InputTypeKind.Enum,
-        Name: enumName,
-        EnumValueType: enumValueType, //EnumValueType and  AllowedValues should be the first field after id and name, so that it can be corrected serialized.
-        AllowedValues: allowValues,
-        Namespace: literalContext.Namespace,
-        Accessibility: undefined,
-        Deprecated: undefined,
-        Description: `The ${enumName}`, // TODO -- what should we put here?
-        IsExtensible: true,
-        IsNullable: false,
-        Usage: "None", // will be updated later
-      };
-      return enumType;
-    }
-  }
-
-  function getInputTypeForEnum(e: Enum, addToCollection: boolean = true): InputEnumType {
-    const name = getTypeName(context, e);
-    let enumType = enums.get(name);
-
-    if (enumType) return enumType;
-
-    // if it's in TCGC model cache, then construct from TCGC
-    if (context.modelsMap?.has(e)) {
-      return fromSdkEnumType(
-        context.modelsMap!.get(e) as SdkEnumType,
-        context,
-        enums,
-        addToCollection
-      );
-    }
-
-    const createdSdkEnumType = getClientType(context, e) as SdkEnumType;
-    context.modelsMap!.set(e, createdSdkEnumType);
-    enumType = fromSdkEnumType(createdSdkEnumType, context, enums);
-    if (addToCollection) enums.set(name, enumType);
-
-    return enumType;
-  }
-
-  function getInputTypeForArray(elementType: Type): InputListType {
-    return {
-      Kind: InputTypeKind.Array,
-      Name: InputTypeKind.Array,
-      ElementType: getInputType(context, getFormattedType(program, elementType), models, enums),
-      IsNullable: false,
-    };
-  }
-
-  function getInputTypeForMap(key: Type, value: Type): InputDictionaryType {
-    return {
-      Kind: InputTypeKind.Dictionary,
-      Name: InputTypeKind.Dictionary,
-      KeyType: getInputType(context, getFormattedType(program, key), models, enums),
-      ValueType: getInputType(context, getFormattedType(program, value), models, enums),
-      IsNullable: false,
-    };
-  }
-
-  function getInputModelForModel(m: Model): InputModelType {
-    if (context.modelsMap!.has(m)) {
-      return fromSdkModelType(context.modelsMap!.get(m) as SdkModelType, context, models, enums);
-    }
-    const createdSdkModelType = getClientType(context, m) as SdkModelType;
-    context.modelsMap!.set(m, createdSdkModelType);
-    return fromSdkModelType(createdSdkModelType, context, models, enums);
-  }
-
-  function getInputModelForIntrinsicType(type: IntrinsicType): InputIntrinsicType {
-    switch (type.name) {
-      case "unknown":
-        return {
-          Kind: InputTypeKind.Intrinsic,
-          Name: InputIntrinsicTypeKind.Unknown,
-          IsNullable: false,
-        } as InputIntrinsicType;
-      case "null":
-        return {
-          Kind: InputTypeKind.Intrinsic,
-          Name: InputIntrinsicTypeKind.Null,
-          IsNullable: false,
-        } as InputIntrinsicType;
-      default:
-        throw new Error(`Unsupported type ${type.name}`);
-    }
-  }
-
-  function getInputTypeForUnion(union: Union): InputUnionType | InputType {
-    const clientType = getClientType(context, union);
-    if (clientType.kind === "enum" && clientType.isFixed === false) {
-      return fromSdkEnumType(clientType, context, enums);
-    }
-
-    let itemTypes: InputType[] = [];
-    const variants = Array.from(union.variants.values());
-
-    let hasNullType = false;
-    for (const variant of variants) {
-      const inputType = getInputType(
-        context,
-        getFormattedType(program, variant.type),
-        models,
-        enums
-      );
-      if (isInputIntrinsicType(inputType) && inputType.Name === InputIntrinsicTypeKind.Null) {
-        hasNullType = true;
-        continue;
-      }
-      itemTypes.push(inputType);
-    }
-
-    if (hasNullType) {
-      itemTypes = itemTypes.map((i) => {
-        i.IsNullable = true;
-        return i;
-      });
-    }
-
-    return itemTypes.length > 1
-      ? {
-          Kind: InputTypeKind.Union,
-          Name: InputTypeKind.Union,
-          UnionItemTypes: itemTypes,
-          IsNullable: false,
-        }
-      : itemTypes[0];
-  }
+  const sdkType = getClientType(context, type, operation);
+  return fromSdkType(sdkType, context, models, enums, literalTypeContext);
 }
 
 export function getUsages(
@@ -567,7 +139,7 @@ export function getUsages(
   for (const type of usages.types) {
     let typeName = "";
     if ("name" in type) typeName = type.name ?? "";
-    const effectiveType = type;
+    let effectiveType = type;
     if (type.kind === "Enum") {
       typeName = getTypeName(context, type);
     }
@@ -575,7 +147,7 @@ export function getUsages(
       typeName = getTypeName(context, effectiveType as Model);
     }
     if (type.kind === "Union") {
-      const clientType = getClientType(context, type);
+      let clientType = getClientType(context, type); // TODO -- we should also pass in an operation as well
       if (clientType.kind === "enum" && clientType.isFixed === false) {
         typeName = clientType.name;
       }
@@ -601,8 +173,9 @@ export function getUsages(
   }
 
   for (const op of ops) {
+    const resourceOperation = getResourceOperation(program, op.operation);
     if (!op.parameters.body?.parameter && op.parameters.body?.type) {
-      let effectiveBodyType = undefined;
+      var effectiveBodyType = undefined;
       const affectTypes: Set<string> = new Set<string>();
       effectiveBodyType = getEffectiveSchemaType(context, op.parameters.body.type);
       if (effectiveBodyType.kind === "Model") {
@@ -667,7 +240,7 @@ export function getUsages(
     // iterate all models to find if it contains literal type properties
     for (const [name, model] of modelMap) {
       // get the usage of this model
-      const usage = usagesMap.get(name);
+      let usage = usagesMap.get(name);
       for (const prop of model.Properties) {
         const type = prop.Type;
         if (!isInputLiteralType(type)) continue;
@@ -736,22 +309,6 @@ export function getUsages(
   }
 }
 
-export function getFormattedType(program: Program, type: Type): FormattedType {
-  let targetType = type;
-  const format = getFormat(program, type);
-  if (type.kind === "ModelProperty") {
-    targetType = type.type;
-  }
-  const encodeData =
-    type.kind === "Scalar" || type.kind === "ModelProperty" ? getEncode(program, type) : undefined;
-
-  return {
-    type: targetType,
-    format: format,
-    encode: encodeData,
-  };
-}
-
 export function navigateModels(
   context: SdkContext<NetEmitterOptions>,
   models: Map<string, InputModelType>,
@@ -762,13 +319,4 @@ export function navigateModels(
       ? fromSdkModelType(model, context, models, enums)
       : fromSdkEnumType(model, context, enums)
   );
-}
-
-// TODO: we should try to remove this when we adopt getAllOperations
-// we should avoid handling raw type definitions because they could be not correctly projected
-// in the given api version
-function getRealType(type: Type, context: SdkContext<NetEmitterOptions>): Type {
-  if ("projector" in context.program)
-    return (context.program as ProjectedProgram).projector.projectedTypes.get(type) ?? type;
-  return type;
 }
