--- packages/http-client-csharp/emitter/src/type/formatted-type.ts
@@ -1,9 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for license information.
-
-import { EncodeData, Type } from "@typespec/compiler";
-export interface FormattedType {
-  type: Type;
-  format?: string;
-  encode?: EncodeData;
-}
