+++ packages/http-client-csharp/emitter/src/type/input-primitive-type-kind.ts
@@ -23,6 +23,7 @@ export enum InputPrimitiveTypeKind {
   DurationConstant = "DurationConstant",
   DurationSeconds = "DurationSeconds",
   DurationSecondsFloat = "DurationSecondsFloat",
+  DurationSecondsDouble = "DurationSecondsDouble",
   ETag = "Etag",
   Float32 = "Float32",
   Float64 = "Float64",
