+++ packages/http-client-csharp/emitter/test/Unit/model-type.test.ts
@@ -1,5 +1,4 @@
 import { TestHost } from "@typespec/compiler/testing";
-import { getAllHttpServices } from "@typespec/http";
 import assert, { deepStrictEqual, strictEqual } from "assert";
 import { beforeEach, describe, it } from "vitest";
 import { createModel } from "../../src/lib/client-model-builder.js";
@@ -7,12 +6,11 @@ import { InputIntrinsicTypeKind } from "../../src/type/input-intrinsic-type-kind
 import { InputModelProperty } from "../../src/type/input-model-property.js";
 import { InputPrimitiveTypeKind } from "../../src/type/input-primitive-type-kind.js";
 import { InputTypeKind } from "../../src/type/input-type-kind.js";
-import { InputDictionaryType, InputEnumType, InputModelType } from "../../src/type/input-type.js";
+import { InputDictionaryType } from "../../src/type/input-type.js";
 import {
   createEmitterContext,
   createEmitterTestHost,
   createNetSdkContext,
-  navigateModels,
   typeSpecCompile,
 } from "./utils/test-util.js";
 
@@ -148,11 +146,9 @@ op test(@body input: Pet): Pet;
     );
     const context = createEmitterContext(program);
     const sdkContext = createNetSdkContext(context);
-    const [services] = getAllHttpServices(program);
-    const modelMap = new Map<string, InputModelType>();
-    const enumMap = new Map<string, InputEnumType>();
-    navigateModels(sdkContext, services[0].namespace, modelMap, enumMap);
-    const pet = modelMap.get("Pet");
+    const codeModel = createModel(sdkContext);
+    const models = codeModel.Models;
+    const pet = models.find((m) => m.Name === "Pet");
     assert(pet !== undefined);
     // assert the discriminator property name
     strictEqual("kind", pet?.DiscriminatorPropertyName);
@@ -197,7 +193,7 @@ op test(@body input: Pet): Pet;
     );
 
     // verify derived model Cat
-    const cat = modelMap.get("Cat");
+    const cat = models.find((m) => m.Name === "Cat");
     assert(cat !== undefined);
     assert(cat.DiscriminatorValue === "Cat");
     assert(cat.BaseModel === pet);
@@ -216,7 +212,7 @@ op test(@body input: Pet): Pet;
     );
 
     // verify derived model Dog
-    const dog = modelMap.get("Dog");
+    const dog = models.find((m) => m.Name === "Dog");
     assert(dog !== undefined);
     assert(dog.DiscriminatorValue === "Dog");
     assert(dog.BaseModel === pet);
@@ -274,11 +270,9 @@ op test(@body input: Pet): Pet;
     );
     const context = createEmitterContext(program);
     const sdkContext = createNetSdkContext(context);
-    const [services] = getAllHttpServices(program);
-    const modelMap = new Map<string, InputModelType>();
-    const enumMap = new Map<string, InputEnumType>();
-    navigateModels(sdkContext, services[0].namespace, modelMap, enumMap);
-    const pet = modelMap.get("Pet");
+    const codeModel = createModel(sdkContext);
+    const models = codeModel.Models;
+    const pet = models.find((m) => m.Name === "Pet");
     assert(pet !== undefined);
     // assert the discriminator property name
     strictEqual("kind", pet?.DiscriminatorPropertyName);
@@ -323,7 +317,7 @@ op test(@body input: Pet): Pet;
     );
 
     // verify derived model Cat
-    const cat = modelMap.get("Cat");
+    const cat = models.find((m) => m.Name === "Cat");
     assert(cat !== undefined);
     assert(cat.DiscriminatorValue === "cat");
     assert(cat.BaseModel === pet);
@@ -342,7 +336,7 @@ op test(@body input: Pet): Pet;
     );
 
     // verify derived model Dog
-    const dog = modelMap.get("Dog");
+    const dog = models.find((m) => m.Name === "Dog");
     assert(dog !== undefined);
     assert(dog.DiscriminatorValue === "dog");
     assert(dog.BaseModel === pet);
